chore: move actions editor to creator
