+++ packages/haiku-creator/index.html
@@ -11,6 +11,7 @@
   <link rel="stylesheet" href="./public/stylesheets/codemirror-haiku.css">
   <link rel="stylesheet" href="./public/stylesheets/icons.css">
   <link rel="stylesheet" href="./public/stylesheets/resizer.css">
+  <link rel="stylesheet" type="text/css" href="../../node_modules/react-selectize/dist/index.css">
 </head>
 <body style="position:relative;">
   <div id="mount" style="width: 100%; height: 100%; overflow: hidden;"></div>
