+++ packages/haiku-glass/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="./public/global.css">
     <link rel="stylesheet" href="./public/icons.css">
-    <link rel="stylesheet" type="text/css" href="../../node_modules/react-selectize/dist/index.css">
     <script>
       window.eval = global.eval = function () {
         // noop: eval is forbidden
