+++ packages/haiku-glass/package.json
@@ -24,7 +24,6 @@
     "raven-js": "^3.18.1",
     "react": "15.6.2",
     "react-dom": "15.6.2",
-    "react-selectize": "^3.0.1",
     "react-transition-group": "^1.1.2"
   },
   "devDependencies": {
